+++ integration/tests/astro/components.test.ts
@@ -1,20 +1,24 @@
 import { expect, test } from '@playwright/test';
 
-import type { FakeUser } from '../../testUtils';
+import type { FakeOrganization, FakeUser } from '../../testUtils';
 import { createTestUtils, testAgainstRunningApps } from '../../testUtils';
 
 testAgainstRunningApps({ withPattern: ['astro.node.withCustomRoles'] })('basic flows for @astro', ({ app }) => {
   test.describe.configure({ mode: 'serial' });
-  let fakeUser: FakeUser;
+
+  let fakeAdmin: FakeUser;
+  let fakeOrganization: FakeOrganization;
 
   test.beforeAll(async () => {
     const m = createTestUtils({ app });
-    fakeUser = m.services.users.createFakeUser();
-    await m.services.users.createBapiUser(fakeUser);
+    fakeAdmin = m.services.users.createFakeUser();
+    const admin = await m.services.users.createBapiUser(fakeAdmin);
+    fakeOrganization = await m.services.users.createFakeOrganization(admin.id);
   });
 
   test.afterAll(async () => {
-    await fakeUser.deleteIfExists();
+    await fakeOrganization.delete();
+    await fakeAdmin.deleteIfExists();
     await app.teardown();
   });
 
@@ -34,11 +38,11 @@ testAgainstRunningApps({ withPattern: ['astro.node.withCustomRoles'] })('basic f
     await u.page.goToRelative('/sign-in');
     await u.po.signIn.waitForMounted();
 
-    await u.po.signIn.setIdentifier(fakeUser.email);
+    await u.po.signIn.setIdentifier(fakeAdmin.email);
     await u.po.signIn.continue();
     await u.page.waitForURL(`${app.serverUrl}/sign-in#/factor-one`);
 
-    await u.po.signIn.setPassword(fakeUser.password);
+    await u.po.signIn.setPassword(fakeAdmin.password);
     await u.po.signIn.continue();
 
     await u.po.expect.toBeSignedIn();
@@ -48,7 +52,7 @@ testAgainstRunningApps({ withPattern: ['astro.node.withCustomRoles'] })('basic f
     const u = createTestUtils({ app, page, context });
     await u.page.goToRelative('/sign-in');
     await u.po.signIn.waitForMounted();
-    await u.po.signIn.signInWithEmailAndInstantPassword({ email: fakeUser.email, password: fakeUser.password });
+    await u.po.signIn.signInWithEmailAndInstantPassword({ email: fakeAdmin.email, password: fakeAdmin.password });
     await u.page.waitForAppUrl('/');
     await u.po.expect.toBeSignedIn();
 
@@ -68,14 +72,14 @@ testAgainstRunningApps({ withPattern: ['astro.node.withCustomRoles'] })('basic f
     const u = createTestUtils({ app, page, context });
     await u.page.goToRelative('/sign-in');
     await u.po.signIn.waitForMounted();
-    await u.po.signIn.signInWithEmailAndInstantPassword({ email: fakeUser.email, password: fakeUser.password });
+    await u.po.signIn.signInWithEmailAndInstantPassword({ email: fakeAdmin.email, password: fakeAdmin.password });
     await u.po.expect.toBeSignedIn();
     await u.po.userButton.waitForMounted();
     await u.page.goToRelative('/user');
     await u.po.userProfile.waitForMounted();
 
     // Streams data from Astro.locals.currentUser()
-    await expect(u.page.getByText(`"firstName":"${fakeUser.firstName}"`)).toBeVisible();
+    await expect(u.page.getByText(`"firstName":"${fakeAdmin.firstName}"`)).toBeVisible();
   });
 
   test('redirects to sign-in when unauthenticated', async ({ page, context }) => {
@@ -91,7 +95,7 @@ testAgainstRunningApps({ withPattern: ['astro.node.withCustomRoles'] })('basic f
     await expect(u.page.getByText('Go to this page to log in')).toBeVisible();
     await u.page.goToRelative('/sign-in');
     await u.po.signIn.waitForMounted();
-    await u.po.signIn.signInWithEmailAndInstantPassword({ email: fakeUser.email, password: fakeUser.password });
+    await u.po.signIn.signInWithEmailAndInstantPassword({ email: fakeAdmin.email, password: fakeAdmin.password });
     await u.po.expect.toBeSignedIn();
     await expect(u.page.getByText('Go to this page to see your profile')).toBeVisible();
     await expect(u.page.getByText('Sign out!')).toBeVisible();
@@ -107,7 +111,7 @@ testAgainstRunningApps({ withPattern: ['astro.node.withCustomRoles'] })('basic f
     await u.page.getByRole('button', { name: /Sign in/i }).click();
 
     await u.po.signIn.waitForMounted();
-    await u.po.signIn.signInWithEmailAndInstantPassword({ email: fakeUser.email, password: fakeUser.password });
+    await u.po.signIn.signInWithEmailAndInstantPassword({ email: fakeAdmin.email, password: fakeAdmin.password });
 
     await u.page.waitForAppUrl('/user');
 
@@ -116,7 +120,7 @@ testAgainstRunningApps({ withPattern: ['astro.node.withCustomRoles'] })('basic f
 
   test('SignUpButton renders and respects props', async ({ page, context }) => {
     const u = createTestUtils({ app, page, context });
-    const fakeUser = u.services.users.createFakeUser({
+    const fakeAdmin = u.services.users.createFakeUser({
       fictionalEmail: true,
       withPhoneNumber: true,
       withUsername: true,
@@ -131,8 +135,8 @@ testAgainstRunningApps({ withPattern: ['astro.node.withCustomRoles'] })('basic f
 
     // Fill in sign up form
     await u.po.signUp.signUpWithEmailAndPassword({
-      email: fakeUser.email,
-      password: fakeUser.password,
+      email: fakeAdmin.email,
+      password: fakeAdmin.password,
     });
 
     // Verify email
@@ -143,7 +147,7 @@ testAgainstRunningApps({ withPattern: ['astro.node.withCustomRoles'] })('basic f
     // Check if user is signed in
     await u.po.expect.toBeSignedIn();
 
-    await fakeUser.deleteIfExists();
+    await fakeAdmin.deleteIfExists();
   });
 
   test('test updateClerkOptions by changing localization on the fly', async ({ page, context }) => {
@@ -160,5 +164,105 @@ testAgainstRunningApps({ withPattern: ['astro.node.withCustomRoles'] })('basic f
     await selectElement.selectOption({ label: 'French' });
 
     await expect(u.page.getByText('pour continuer vers')).toBeVisible();
+  });
+
+  // ---- react/protect
+  test('only admin react', async ({ page, context }) => {
+    const u = createTestUtils({ app, page, context });
+    await u.page.goToRelative('/react/sign-in#/?redirect_url=/react');
+    await u.po.signIn.waitForMounted();
+    await u.po.signIn.signInWithEmailAndInstantPassword({ email: fakeAdmin.email, password: fakeAdmin.password });
+    await u.po.expect.toBeSignedIn();
+    await u.po.organizationSwitcher.waitForMounted();
+    await u.po.organizationSwitcher.toggleTrigger();
+    await u.page.locator('.cl-organizationSwitcherPreviewButton').click();
+    await u.po.organizationSwitcher.waitForAnOrganizationToSelected();
+    await u.page.goToRelative('/react/only-admins');
+    await expect(u.page.getByText("I'm an admin")).toBeVisible();
+  });
+
+  test('only member react', async ({ page, context }) => {
+    const u = createTestUtils({ app, page, context });
+    await u.page.goToRelative('/react/sign-in#/?redirect_url=/react/only-members');
+    await u.po.signIn.waitForMounted();
+    await u.po.signIn.signInWithEmailAndInstantPassword({ email: fakeAdmin.email, password: fakeAdmin.password });
+    await u.po.expect.toBeSignedIn();
+    await expect(u.page.getByText('Not a member')).toBeVisible();
+  });
+
+  // --- react/components
+  test('react/ render user button', async ({ page, context }) => {
+    const u = createTestUtils({ app, page, context });
+    await u.page.goToRelative('/react/sign-in#/?redirect_url=/react');
+    await u.po.signIn.waitForMounted();
+    await u.po.signIn.signInWithEmailAndInstantPassword({ email: fakeAdmin.email, password: fakeAdmin.password });
+    await u.page.waitForAppUrl('/react');
+    await u.po.expect.toBeSignedIn();
+
+    await u.po.userButton.waitForMounted();
+    await u.po.userButton.toggleTrigger();
+    await u.po.userButton.waitForPopover();
+
+    await u.po.userButton.toHaveVisibleMenuItems([/Manage account/i, /Sign out$/i]);
+
+    await u.po.userButton.triggerManageAccount();
+    await u.po.userProfile.waitForUserProfileModal();
+
+    await expect(u.page.getByText(/profile details/i)).toBeVisible();
+  });
+
+  test('react/ render user profile with streamed data', async ({ page, context }) => {
+    const u = createTestUtils({ app, page, context });
+    await u.page.goToRelative('/react/sign-in#/?redirect_url=/react');
+    await u.po.signIn.waitForMounted();
+    await u.po.signIn.signInWithEmailAndInstantPassword({ email: fakeAdmin.email, password: fakeAdmin.password });
+    await u.po.expect.toBeSignedIn();
+    await u.po.userButton.waitForMounted();
+    await u.page.goToRelative('/react/user');
+    await u.po.userProfile.waitForMounted();
+
+    await expect(u.page.getByText(`My name is: ${fakeAdmin.firstName}`)).toBeVisible();
+
+    // Streams data from Astro.locals.currentUser()
+    await expect(u.page.getByText(`"firstName":"${fakeAdmin.firstName}"`)).toBeVisible();
+  });
+
+  test('react/ SignedIn, SignedOut SSR', async ({ page, context }) => {
+    const u = createTestUtils({ app, page, context });
+    await u.page.goToRelative('/react');
+    await expect(u.page.getByText('Go to this page to log in')).toBeVisible();
+    await u.page.goToRelative('/react/sign-in#/?redirect_url=/react');
+    await u.po.signIn.waitForMounted();
+    await u.po.signIn.signInWithEmailAndInstantPassword({ email: fakeAdmin.email, password: fakeAdmin.password });
+    await u.po.expect.toBeSignedIn();
+    await expect(u.page.getByText('Go to this page to see your profile')).toBeVisible();
+    await expect(u.page.getByText('Sign out!')).toBeVisible();
+  });
+
+  test('react/ render content based on Clerk loaded status', async ({ page, context }) => {
+    const u = createTestUtils({ app, page, context });
+    await u.page.goToRelative('/utility');
+    await expect(u.page.getByText('Clerk is loading')).toBeVisible();
+    await expect(u.page.getByText('Clerk is loaded')).toBeHidden();
+    await u.page.waitForClerkJsLoaded();
+    await expect(u.page.getByText('Clerk is loaded')).toBeVisible();
+    await expect(u.page.getByText('Clerk is loading')).toBeHidden();
+  });
+
+  // ----- redirect
+  test('redirects to sign-in when unauthenticated (middleware)', async ({ page, context }) => {
+    const u = createTestUtils({ app, page, context });
+    await u.page.goToAppHome();
+    await u.page.getByRole('link', { name: 'User', exact: true }).click();
+    await u.page.waitForURL(`${app.serverUrl}/sign-in?redirect_url=${encodeURIComponent(`${app.serverUrl}/user`)}`);
+  });
+
+  test('redirects to sign-in when unauthenticated (page)', async ({ page, context }) => {
+    const u = createTestUtils({ app, page, context });
+    await u.page.goToAppHome();
+    await u.page.getByRole('link', { name: 'Organization', exact: true }).click();
+    await u.page.waitForURL(
+      `${app.serverUrl}/sign-in?redirect_url=${encodeURIComponent(`${app.serverUrl}/organization`)}`,
+    );
   });
 });
